+++ src/datasources/test-plans/components/TestPlansVariableQueryEditor.test.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { render, waitFor } from '@testing-library/react';
+import { act, render, waitFor } from '@testing-library/react';
 import { TestPlansVariableQueryEditor } from './TestPlansVariableQueryEditor';
 import { QueryEditorProps } from '@grafana/data';
 import { TestPlansDataSource } from '../TestPlansDataSource';
@@ -11,6 +11,22 @@ const mockOnChange = jest.fn();
 const mockOnRunQuery = jest.fn();
 const mockDatasource = {
   prepareQuery: jest.fn((query: TestPlansVariableQuery) => query),
+  workspaceUtils: {
+    getWorkspaces: jest.fn().mockResolvedValue(
+      new Map([
+        ['1', { id: '1', name: 'WorkspaceName' }],
+        ['2', { id: '2', name: 'AnotherWorkspaceName' }],
+      ])
+    ),
+  },
+  systemUtils: {
+    getSystemAliases: jest.fn().mockResolvedValue(
+      new Map([
+        ['1', { id: '1', alias: 'System 1' }],
+        ['2', { id: '2', alias: 'System 2' }],
+      ])
+    ),
+  }
 } as unknown as TestPlansDataSource;
 
 const defaultProps: QueryEditorProps<TestPlansDataSource, TestPlansVariableQuery> = {
@@ -27,13 +43,15 @@ describe('TestPlansVariableQueryEditor', () => {
     jest.clearAllMocks();
   });
 
-  function renderElement(query: TestPlansVariableQuery = { refId: 'A' }) {
-    const reactNode = React.createElement(TestPlansVariableQueryEditor, { ...defaultProps, query });
-    return render(reactNode);
+  async function renderElement(query: TestPlansVariableQuery = { refId: 'A' }) {
+    return await act(async () => {
+      const reactNode = React.createElement(TestPlansVariableQueryEditor, { ...defaultProps, query });
+      return render(reactNode);
+    });
   }
 
   it('should render default query', async () => {
-    const container = renderElement();
+    const container = await renderElement();
 
     await waitFor(() => {
       const orderBy = container.getAllByRole('combobox')[0];
@@ -55,7 +73,7 @@ describe('TestPlansVariableQueryEditor', () => {
   });
 
   it('only allows numbers in Take field', async () => {
-    const container = renderElement();
+    const container = await renderElement();
 
     const recordCountInput = container.getByRole('spinbutton');
 
@@ -74,9 +92,33 @@ describe('TestPlansVariableQueryEditor', () => {
     });
   });
 
+  it('should load workspaces and set them in state', async () => {
+    await renderElement();
+
+    expect(mockDatasource.workspaceUtils.getWorkspaces()).toBeDefined();
+    await expect(mockDatasource.workspaceUtils.getWorkspaces()).resolves.toEqual(
+      new Map([
+        ['1', { id: '1', name: 'WorkspaceName' }],
+        ['2', { id: '2', name: 'AnotherWorkspaceName' }],
+      ])
+    );
+  });
+
+  it('should load system names', async () => {
+    await renderElement();
+    const result = await mockDatasource.systemUtils.getSystemAliases();
+    expect(result).toBeDefined();
+    expect(result).toEqual(
+      new Map([
+        ['1', { id: '1', alias: 'System 1' }],
+        ['2', { id: '2', alias: 'System 2' }],
+      ])
+    );
+  });
+
   describe('onChange', () => {
     it('should call onChange with order by when user selects order by', async () => {
-      const container = renderElement();
+      const container = await renderElement();
       const orderBySelect = container.getAllByRole('combobox')[0];
 
       userEvent.click(orderBySelect);
@@ -88,7 +130,7 @@ describe('TestPlansVariableQueryEditor', () => {
     });
 
     it('should call onChange with descending when user toggles descending', async () => {
-      const container = renderElement();
+      const container = await renderElement();
       const descendingCheckbox = container.getByRole('checkbox');
 
       userEvent.click(descendingCheckbox);
@@ -99,7 +141,7 @@ describe('TestPlansVariableQueryEditor', () => {
     });
 
     it('should call onChange with record count when user enters record count', async () => {
-      const container = renderElement();
+      const container = await renderElement();
       const recordCountInput = container.getByRole('spinbutton');
 
       await userEvent.clear(recordCountInput);
@@ -112,7 +154,7 @@ describe('TestPlansVariableQueryEditor', () => {
     });
 
     it('should call onChange when query by changes', async () => {
-      const container = renderElement();
+      const container = await renderElement();
 
       const queryBuilder = container.getByRole('dialog');
       expect(queryBuilder).toBeInTheDocument();
@@ -127,7 +169,7 @@ describe('TestPlansVariableQueryEditor', () => {
     });
 
     it('should show error message when when user changes take to number greater than max take', async () => {
-      const container = renderElement();
+      const container = await renderElement();
       const takeInput = container.getByRole('spinbutton');
       mockOnChange.mockClear();
 
@@ -142,7 +184,7 @@ describe('TestPlansVariableQueryEditor', () => {
     });
 
     it('should show error message when when user changes take to number less than min take', async () => {
-      const container = renderElement();
+      const container = await renderElement();
       const takeInput = container.getByRole('spinbutton');
       mockOnChange.mockClear();
 
@@ -156,7 +198,7 @@ describe('TestPlansVariableQueryEditor', () => {
     });
 
     it('should not show error message when when user changes take to number between min and max take', async () => {
-      const container = renderElement();
+      const container = await renderElement();
       const takeInput = container.getByRole('spinbutton');
 
       // User enters a value greater than max take