+++ src/datasources/test-plans/components/TestPlansVariableQueryEditor.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useState } from 'react';
+import React, { useCallback, useEffect, useState } from 'react';
 import { QueryEditorProps, SelectableValue } from '@grafana/data';
 import { OrderBy, TestPlansVariableQuery } from '../types';
 import { AutoSizeInput, InlineField, InlineSwitch, Select, VerticalGroup } from '@grafana/ui';
@@ -6,6 +6,8 @@ import { validateNumericInput } from 'core/utils';
 import { TestPlansDataSource } from '../TestPlansDataSource';
 import { TestPlansQueryBuilder } from './query-builder/TestPlansQueryBuilder';
 import { recordCountErrorMessages, TAKE_LIMIT } from '../constants/QueryEditor.constants';
+import { Workspace } from 'core/types';
+import { SystemAlias } from 'shared/types/QuerySystems.types';
 
 type Props = QueryEditorProps<TestPlansDataSource, TestPlansVariableQuery>;
 
@@ -13,6 +15,24 @@ export function TestPlansVariableQueryEditor({ query, onChange, datasource }: Pr
   query = datasource.prepareQuery(query);
   const [recordCountInvalidMessage, setRecordCountInvalidMessage] = useState<string>('');
 
+  const [workspaces, setWorkspaces] = useState<Workspace[] | null>(null);
+  const [systemAliases, setSystemAliases] = useState<SystemAlias[] | null>(null);
+
+  useEffect(() => {
+    const loadWorkspaces = async () => {
+      const workspaces = await datasource.workspaceUtils.getWorkspaces();
+      setWorkspaces(Array.from(workspaces.values()));
+    };
+
+    loadWorkspaces();
+
+    const loadSystemAliases = async () => {
+      const systemAliases = await datasource.systemUtils.getSystemAliases();
+      setSystemAliases(Array.from(systemAliases.values()));
+    };
+
+    loadSystemAliases();
+  }, [datasource]);
 
   const handleQueryChange = useCallback(
     (query: TestPlansVariableQuery): void => {
@@ -30,17 +50,13 @@ export function TestPlansVariableQueryEditor({ query, onChange, datasource }: Pr
 
   const recordCountChange = (event: React.FormEvent<HTMLInputElement>) => {
     const value = parseInt((event.target as HTMLInputElement).value, 10);
-    switch (true) {
-      case isNaN(value) || value < 0:
-        setRecordCountInvalidMessage(recordCountErrorMessages.greaterOrEqualToZero);
-        break;
-      case value > TAKE_LIMIT:
-        setRecordCountInvalidMessage(recordCountErrorMessages.lessOrEqualToTenThousand);
-        break;
-      default:
-        setRecordCountInvalidMessage('');
-        handleQueryChange({ ...query, recordCount: value });
-        break;
+    if (isNaN(value) || value < 0) {
+      setRecordCountInvalidMessage(recordCountErrorMessages.greaterOrEqualToZero);
+    } else if (value > TAKE_LIMIT) {
+      setRecordCountInvalidMessage(recordCountErrorMessages.lessOrEqualToTenThousand);
+    } else {
+      setRecordCountInvalidMessage('');
+      handleQueryChange({ ...query, recordCount: value });
     }
   };
 
@@ -56,6 +72,8 @@ export function TestPlansVariableQueryEditor({ query, onChange, datasource }: Pr
       <InlineField label="Query By" labelWidth={25} tooltip={tooltips.queryBy}>
         <TestPlansQueryBuilder
           filter={query.queryBy}
+          workspaces={workspaces}
+          systemAliases={systemAliases}
           globalVariableOptions={[]}
           onChange={(event: any) => onQueryByChange(event.detail.linq)}
         ></TestPlansQueryBuilder>